Address feedback
- nicer scripts
- cleaner code
- catch attempted uploads of files > 128MB
- time-based upload file names
- use IOptions to hold fileshare info
- mention

+++ FileShares/scripts/remove-user-and-share.ps1
@@ -4,18 +4,48 @@ Param(
     [string]$UserName = "shareWriteUser"
 )
 $ErrorActionPreference = "Stop"
+#Requires -RunAsAdministrator
+#Requires -Modules Microsoft.PowerShell.LocalAccounts, SmbShare
+Import-Module Microsoft.PowerShell.LocalAccounts -SkipEditionCheck
 
-Remove-SmbShare -Name $ShareName
-Write-Host "SMB share removed."
+if (Get-SmbShare $ShareName -ErrorAction SilentlyContinue)
+{
+    Remove-SmbShare -Name $ShareName
+    Write-Host "SMB share $ShareName removed."
+}
+else
+{
+    Write-Host "SMB share $ShareName was not found."
+}
 
-Write-Host "Removing $UserName from local 'Users' group..."
-Remove-LocalGroupMember -Group "Users" -Member $UserName
-Write-Host "User removed from group."
+if (Get-LocalUser -Name $UserName -ErrorAction SilentlyContinue)
+{
+    if (Get-LocalGroupMember -Group "Users" -Member $UserName -ErrorAction SilentlyContinue)
+    {
+        Write-Host "Removing $UserName from local 'Users' group..."
+        Remove-LocalGroupMember -Group "Users" -Member $UserName
+        Write-Host "User removed from group."
+    }
+    else
+    {
+        Write-Host "User $UserName was not found in 'Users' group."
+    }
+    Write-Host "Removing local user $UserName..."
+    Remove-LocalUser -Name $UserName
+    Write-Host "User completely removed."
+}
+else
+{
+    Write-Host "User $UserName was not found."
+}
 
-Write-Host "Removing local user $UserName..."
-Remove-LocalUser -Name $UserName
-Write-Host "User completely removed."
-
-Write-Host "Removing $FolderPath from disk..."
-Remove-Item -Path $FolderPath -Recurse
-Write-Host "Folder completely removed."
+if (Get-Item -Path $FolderPath -ErrorAction SilentlyContinue)
+{
+    Write-Host "Removing $FolderPath from disk..."
+    Remove-Item -Path $FolderPath -Recurse
+    Write-Host "Directory completely removed."
+}
+else
+{
+    Write-Host "$FolderPath was not found."
+}
